+++ src/components/Footer/styles.module.scss
@@ -1,0 +1,24 @@
+@import "includes";
+
+.container {
+  border-top: 2px solid var(--tertiary-color);
+  display: flex;
+  justify-content: space-between;
+  padding-top: 2rem;
+}
+
+.content {
+  display: flex;
+  flex-direction: column;
+}
+
+.contact {
+  color: var(--tertiary-color);
+  font-size: 1.2rem;
+  line-height: 1.4rem;
+
+  @include fromLarge {
+    font-size: 2.1rem;
+    line-height: 2.8rem;
+  }
+}
