+++ src/components/Home/styles.module.scss
@@ -1,0 +1,11 @@
+@import "includes";
+
+.footer {
+  @include grid-column-span($grid-columns-mobile);
+  @include grid-column-start(1);
+  margin-top: 8rem;
+
+  @include fromLarge {
+    @include grid-column-span($grid-columns-desktop);
+  }
+}
