+++ src/pages/index.js
@@ -1,22 +1,3 @@
-import React from "react"
+import Home from "../components/Home"
 
-import Layout from "../components/Layout"
-import SEO from "../components/seo"
-import HomeHeader from "../components/Header"
-import RestaurantList from "../components/RestaurantList"
-import Footer from "../components/Footer"
-import useTheme from "../hooks/useTheme"
-
-const IndexPage = () => {
-  useTheme("home")
-
-  return (
-    <Layout>
-      <SEO title="Home" />
-      <RestaurantList />
-      <Footer />
-    </Layout>
-  )
-}
-
-export default IndexPage
+export default Home
